Widen chapter indices to u16 for long tables of contents
Books with more than 255 chapters wrapped or saturated everywhere the
chapter index travelled as a u8: the Chapters overview could not select
past entry 255, the reducer clamped the firmware's uncapped
ChapterCursor/Loaded corrections, and a restored position past chapter
255 snapped back.

Chapter indices and counts are now u16 end to end — RenderRequest,
StorageCommand, LibraryEvent, ReaderState, the reader store's loaded
chapter and UI count, the UiShell/colophon renderers, and both
emulators. The persisted AppStateRecord and POS.BIN already stored u16
chapters, so no on-disk format changes. Adds a reducer test driving
selection across the 255/256 boundary.

Ported from upstream 8da2d42.

+++ fw/src/reader_store.rs
@@ -175,7 +175,7 @@ pub(crate) struct ReaderStore {
     active_index: Option<usize>,
     pub(crate) current_index: Option<usize>,
     pub(crate) loaded_index: Option<usize>,
-    pub(crate) loaded_chapter: u8,
+    pub(crate) loaded_chapter: u16,
     pub(crate) reader_status: BookLoadStatus,
     pub(crate) title: String<64>,
     pub(crate) author: String<64>,
@@ -699,7 +699,7 @@ impl ReaderStore {
         self.clear_book_index();
     }
 
-    pub(crate) fn finish_book_load(&mut self, index: usize, chapter: u8, status: BookLoadStatus) {
+    pub(crate) fn finish_book_load(&mut self, index: usize, chapter: u16, status: BookLoadStatus) {
         if matches!(status, BookLoadStatus::Ready) {
             self.set_current_index(index);
         }
@@ -1235,10 +1235,10 @@ impl ReaderStore {
         self.error.as_str()
     }
 
-    pub(crate) fn chapter_count_for_ui(&self) -> u8 {
+    pub(crate) fn chapter_count_for_ui(&self) -> u16 {
         // The full on-disk count (when known) drives the overview's
         // selection range; it can run past the 128-entry resident/event
-        // caps, so only the u8 message width bounds it.
+        // caps, so only the u16 message width bounds it.
         let count = if self.toc_total > 0 {
             self.toc_total
         } else if self.toc_count > 0 {
@@ -1246,7 +1246,7 @@ impl ReaderStore {
         } else {
             self.book_section_count
         };
-        count.min(u8::MAX as usize).max(1) as u8
+        count.min(u16::MAX as usize).max(1) as u16
     }
 
     pub(crate) fn set_current_index(&mut self, index: usize) {
